refactor(agent): 拆分 RootAgentSession 状态机为独立子类
将 root-agent-session.ts 中混在一起的 6 个 RootAgentState 实现拆到
session/states/ 目录，并提取 state.types.ts / state-id.ts /
persistence-normalize.ts 三组纯函数与类型，避免协调器、State 子类、
持久化助手挤在 1429 行的单一文件里。

state 子类通过新增的 RootAgentStateHost 接口访问 session 内部状态，
打破 state ↔ session 的循环依赖，保留原有 public API 不变。
RootAgentSession 主文件从 1429 行降至 809 行。
